Added initial dependency injection, compressed CRUD into a single controller
